proxy/src/test/java/org/fedoraproject/candlepin/pinsetter/tasks/CertificateRevocationListTaskTest.java
/**
 * Copyright (c) 2009 Red Hat, Inc.
 *
 * This software is licensed to you under the GNU General Public License,
 * version 2 (GPLv2). There is NO WARRANTY for this software, express or
 * implied, including the implied warranties of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. You should have received a copy of GPLv2
 * along with this software; if not, see
 * http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
 *
 * Red Hat trademarks are not licensed under GPLv2. No permission is
 * granted to use or replicate Red Hat trademarks that are incorporated
 * in this software or its documentation.
 */
package org.fedoraproject.candlepin.pinsetter.tasks;

import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;

import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import org.fedoraproject.candlepin.test.DatabaseTestFixture;
import org.junit.Before;
import org.junit.Test;
import org.quartz.JobExecutionException;


/**
 * CertificateRevocationListTaskTest
 */
public class CertificateRevocationListTaskTest extends DatabaseTestFixture{
    private final ByteArrayOutputStream out = new ByteArrayOutputStream();
    private final ByteArrayOutputStream err = new ByteArrayOutputStream();
    

    @Before
    public void setupStreams() {
        System.setOut(new PrintStream(out));
        System.setErr(new PrintStream(err));
    }

    @Test
    public void testExecute() throws JobExecutionException {
        CertificateRevocationListTask task = new CertificateRevocationListTask(null);
        task.setEntCertCurator(entCertCurator);
        task.execute(null);
        assertNotNull(out.toString());
        assertTrue(out.toString().startsWith("crl task ran:"));
    }
}
